Implement party HP bar sync (#211)
* feat(Player): Implementation of packetPlayerHpChange to broadcast HP change. should make party hp bar usable

as suggested in issue #180 I've implmented the packet "packetPlayerHpChange" using opcode 0x73 to send updates on HP changes (using giveHP/setHP) functios.

* feat(Party): Implementation of syncPlayersHP

While using setHP to broadcast hp works when updating the value, syncPlayersHP is being used when a player is joined the party. therefore broadcasting the HP of all members of the party is necessary whenever a new player joins.

* docs(README): Marked Party HP Bar as done

---------

Co-authored-by: Hucaru <7339035+Hucaru@users.noreply.github.com>